pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "NoiseDiffusion"
version = "1.2"
description = "Evaluating Text-to-Visual Generation with Image-to-Text Generation."
authors = [
    {name="Zhiqiu Lin", email="zl279@cornell.edu"},
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "ftfy>=6.1.1",
    "tqdm>=4.64.1",
    "gdown>=4.7.1",
    "huggingface-hub>=0.19.4",
    "matplotlib>=3.6.2",
    "numpy>=1.26.2",
    "open-clip-torch>=2.23.0",
    "openai>=1.29.0",
    "opencv-python>=4.6.0.66",
    "opencv-python-headless",
    "pandas>=2.1.4",
    "scipy>=1.11.4",
    "sentencepiece>=0.1.99",
    "transformers==4.36.1",
    "datasets>=2.15.0",
    "tokenizers",
    "omegaconf",
    "iopath",
    "fairscale",
    # for clipscore
    "scikit-learn",
    "pycocoevalcap",
    "image-reward",
    "hpsv2",
    "fire==0.4.0",
    "tiktoken>=0.7.0",
]

[tool.setuptools]
include-package-data = true
packages = ["t2v_metrics", "t2v_metrics.models"]

[tool.setuptools.package-data]
't2v_metrics' = ['**/*.json', '**/*.yaml', '**/*.py']

[project.urls]
Home = "https://linzhiqiu.github.io/papers/vqascore"